requirements.txt
pyyaml==6.0.2
pytest==8.3.3
fastapi==0.115.6
uvicorn==0.32.1
httpx==0.28.1
